+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Changed the brightness sliders, so they have a larger hit test area.
 - The brightness sliders now extend further up.
 - Added a placeholder for German localization.
+- Made it so that the alarm display at the bottom of the Alarm Editor is set to full brightness when editing (it was dimmed, before, if the app was dimmed).
 
 **Version 1.0.0.2016** *(December 18, 2018)*
 - Fixed a minor cosmetic bug, where switching out of the app would not always restore the correct original brightness level.